# Introduce new tracking sequence for phase1 (2017)