--- a/Configuration/StandardSequences/python/Reconstruction_cff.py
+++ b/Configuration/StandardSequences/python/Reconstruction_cff.py
@@ -61,7 +61,7 @@ globalreco_tracking = cms.Sequence(offlineBeamSpot*
                           vertexreco)
 _globalreco_tracking_Phase1PU70 = globalreco_tracking.copy()
 _globalreco_tracking_Phase1PU70.replace(trackingGlobalReco, recopixelvertexing+trackingGlobalReco)
-eras.trackingPhase1.toReplaceWith(globalreco_tracking, _globalreco_tracking_Phase1PU70)
+eras.trackingPhase1PU70.toReplaceWith(globalreco_tracking, _globalreco_tracking_Phase1PU70)
 
 globalreco = cms.Sequence(globalreco_tracking*
                           hcalGlobalRecoSequence*