--- a/RecoPixelVertexing/Configuration/python/RecoPixelVertexing_cff.py
+++ b/RecoPixelVertexing/Configuration/python/RecoPixelVertexing_cff.py
@@ -16,4 +16,4 @@ PixelLayerTripletsPreSplitting = PixelLayerTriplets.clone(
 )
 _recopixelvertexing_Phase1PU70 = recopixelvertexing.copy()
 _recopixelvertexing_Phase1PU70.replace(PixelLayerTriplets, PixelLayerTripletsPreSplitting)
-eras.trackingPhase1.toReplaceWith(recopixelvertexing, _recopixelvertexing_Phase1PU70)
+eras.trackingPhase1PU70.toReplaceWith(recopixelvertexing, _recopixelvertexing_Phase1PU70)