--- a/RecoTracker/FinalTrackSelectors/python/TrackCutClassifier_cff.py
+++ b/RecoTracker/FinalTrackSelectors/python/TrackCutClassifier_cff.py
@@ -1,5 +1,5 @@
 from Configuration.StandardSequences.Eras import eras
 from RecoTracker.FinalTrackSelectors.TrackCutClassifier_cfi import *
-eras.trackingPhase1.toModify(TrackCutClassifier,
+eras.trackingPhase1PU70.toModify(TrackCutClassifier,
     vertices = "pixelVertices"
 )