--- a/RecoTracker/FinalTrackSelectors/python/earlyGeneralTracks_cfi.py
+++ b/RecoTracker/FinalTrackSelectors/python/earlyGeneralTracks_cfi.py
@@ -22,6 +22,33 @@ earlyGeneralTracks.inputClassifiers =["initialStep",
                                       "pixelLessStep",
                                       "tobTecStep"
                                       ]
+eras.trackingPhase1.toModify(
+    earlyGeneralTracks,
+    trackProducers = [
+        'initialStepTracks',
+        'highPtTripletStepTracks',
+        'jetCoreRegionalStepTracks',
+        'lowPtQuadStepTracks',
+        'lowPtTripletStepTracks',
+        'detachedQuadStepTracks',
+        #'detachedTripletStepTracks', # FIXME: disabled for now
+        'mixedTripletStepTracks',
+        'pixelLessStepTracks',
+        'tobTecStepTracks'
+    ],
+    inputClassifiers = [
+        "initialStep",
+        "highPtTripletStep",
+        "jetCoreRegionalStep",
+        "lowPtQuadStep",
+        "lowPtTripletStep",
+        "detachedQuadStep",
+        #"detachedTripletStep", # FIXME: disabled for now
+        "mixedTripletStep",
+        "pixelLessStep",
+        "tobTecStep"
+    ],
+)
 
 # For Phase1PU70
 from RecoTracker.FinalTrackSelectors.trackListMerger_cfi import trackListMerger as _trackListMerger