This PR replaces the "Phase1PU70" tracking with the one presented in Feb 1 TRK POG meeting https://indico.cern.ch/event/489234/contribution/5/attachments/1220760/1784567/slides_phase1.pdf (and in the closed PR #13149). The "Phase1PU70" tracking can still be run with `Run2_2017_trackingPhase1PU70` era, and there should be no changes in it (i.e. `Run2_2017_trackingPhase1PU70` with this PR should give the same tracks as `Run2_2017` without this PR).

Tested in CMSSW_8_1_X_2016-03-15-1100. There should be no changes in <= 2016 workflows, but in 2017 workflows changes are expected in all(?) track-related variables.

about all track-related will change in 2017 workflows.

@rovere @VinInn