@cmsbuild 2016-03-16
A new Pull Request was created by  @makortel (Matti Kortelainen) for CMSSW_8_1_X.

It involves the following packages:

Configuration/StandardSequences
RecoJets/JetProducers
RecoMuon/Configuration
RecoPixelVertexing/Configuration
RecoPixelVertexing/PixelTrackFitting
RecoPixelVertexing/PixelTriplets
RecoTracker/ConversionSeedGenerators
RecoTracker/FinalTrackSelectors
RecoTracker/IterativeTracking
RecoTracker/TkSeedGenerator
RecoTracker/TkSeedingLayers
Validation/RecoTrack


@cvuosalo, @cmsbuild, @franzoni, @deguio, @slava77, @vanbesien, @davidlange6 can you please review it and eventually sign? Thanks.
@ghellwig, @TaiSakuma, @yslai, @abbiendi, @rappoccio, @Martin-Grunewald, @battibass, @ahinzmann, @jhgoh, @jdolen, @cerati, @trocino, @schoef, @rociovilar, @GiacomoSguazzoni, @rovere, @VinInn, @bellan, @nhanvtran, @wmtford, @mschrode, @dgulhan, @gpetruc, @istaslis, @mariadalfonso, @bachtis this is something you requested to watch as well.
@slava77, @degano, @smuzaffar you are the release manager for this.

cms-bot commands are list here cms-sw/cmssw#13028