@makortel 2016-03-16
I have MTV comparison with 1000 TTbar+PU events (GEN-SIM-DIGI-RAW from 8_0_0_patch1 reconstructed in  CMSSW_8_1_X_2016-03-15-1100) in https://mkortela.web.cern.ch/mkortela/tracking/validation/CMSSW_8_1_X_phase1_pr13736/. The color coding matches to the slides mentioned in description, i.e. the expected changes in with `Run2_2017` era are from red to black.