@cmsbuild 2016-03-16
The tests are being triggered in jenkins.