@slava77 2016-03-16
@makortel I noticed this too.
I have alerted @smuzaffar and @degano 
The workflow list is apparently hardcoded in jenkins comparison jobs.