@slava77 2016-03-16
@cmsbuild please test