@cmsbuild 2016-03-16
+1
Tested at: 26bfb299a20c0b2bdad1546baed59db4a816c7d0
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-13736/11928/summary.html