@makortel 2016-03-16
Now the comparisons seem be produced nicely, thanks @smuzaffar.