@bachtis 2016-03-16
Hi . stupid question: Does this file need change?
https://github.com/cms-sw/cmssw/blob/CMSSW_8_1_X/RecoParticleFlow/PFTracking/src/PFTrackAlgoTools.cc