@slava77 2016-03-19
+1

for https://github.com/cms-sw/cmssw/pull/13736	26bfb29
- code changes are in line with the description: more run2-like tracking sequence is now delivered with era.trackingPhase1
- jenkins tests pass and comparisons show differences only in 2017 scenarios.
- tested locally with more events, looking mostly at just 10024 and 10224: there are generally more tracks,  substantially more duplicate tracks (not so good); displacedTrk and displacedMuon tracks are now ~fully efficient; number of TEC clusters on track is up by ~30% (compared to ~5% increase in other strip tracker parts). 
- based on 10224:
    - generalTracks: +16%
    - pf charged hadrons: + 25% (the fraction of good quality tracks is higher)
    - conversionStepTracks count about doubles
    - particleFlowDisplacedVtx count is up by O(10) ( 6 -> 81)
- technical performance from 10224 (ttbar 2017 with PU): about 10% more CPU is spent after the update (mainly increases from tracking)

This is somewhat OK for a work in progress setup.

Some oddities follow:
- mixedTripletStep\* takes ~10times less
- tobTecStepSeeds CPU changes from 94 to 2 ms/evt, while the tobTecStepTrackCandidates changes only from 270 to 240 ms/evt. Why is seeding so much faster?

in a sample without PU (10024) most changes in generalTracks are in prob~0 (as if on average only bad tracks are added):
<img width="871" alt="wf10024_gentk_chi2prob" src="https://cloud.githubusercontent.com/assets/4676718/13901409/2382421c-ede0-11e5-9f6d-8488ec8cc416.png">

In a sample with PU there are more "normal" tracks as well:
<img width="862" alt="wf10224_gentk_chi2prob" src="https://cloud.githubusercontent.com/assets/4676718/13901418/6113cc7c-ede0-11e5-8608-7e10f256714a.png">
Duplicate rate is pretty large .. supposedly to be cleaned up soon
<img width="879" alt="wf10224_gentk_dups_vs_eta" src="https://cloud.githubusercontent.com/assets/4676718/13901419/71cc3568-ede0-11e5-92ae-d83e9a8c4744.png">

Compared to 25202: 
- initialStepSeeds\* take about 6 times more (additional layer of pixels isn't used effectively?) from 60 ms/ev to 370 ms/ev
- detachedTripletStep\* and pixelPairStep\* (total of 1.1s/ev + 1.6 s/ev) are "replaced" by detachedQuadStep\* , highPtTripletStep\*, and lowPtQuadStep (totals of 1.9 s/ev, 1.2 s/ev, and 1.1 s/ev) with the net time increase of 1.5 s/ev. At least now it looks like an extra layer of pixels makes it harder to do tracking. Hopefully this will change.