@makortel 2016-03-20
@slava77 Thanks for the review, please find some answers and further comments below.

> mixedTripletStep* takes ~10times less

This happens because in trackingPhase1PU70 mixedTripletStep* have 12 (A) + 4 (B) seeding layer sets, while in Run2 and trackingPhase1 they have only 2+1 sets, i.e. a reduction by roughly an order of magnitude in the number of seeds is expected.

> tobTecStepSeeds CPU changes from 94 to 2 ms/evt, while the tobTecStepTrackCandidates changes only from 270 to 240 ms/evt. Why is seeding so much faster?

In trackingPhase1PU70 `tobTecStepSeeds` is the seed producer, while in Run2 and trackingPhase1 it is a seed combiner, combining seeds produced by `tobTecStepSeedsTripl` and `tobTecStepSeedsPair`. Including those, the seeding time in trackingPhase1 should roughly double compared to trackingPhase1PU70. Compared to Run 2, the seeding time in trackingPhase1 is roughly the same.

> initialStepSeeds* take about 6 times more (additional layer of pixels isn't used effectively?) from 60 ms/ev to 370 ms/ev

Since you had "*" above I have to ask if you included `initialStepSeedsPreSplitting` there? For apples-to-apples comparison between trackingPhase1 and trackingPhase1PU70 only `initialStepSeeds` should be counted (pre-splitting step should accounted separately for the jet-core pixel cluster splitting). My tests show 2-2.5x increase in the timing of `initialStepSeeds`. This is caused mainly by the increase in the number of layer-triplets from 4 to 14 and that shows up in the timing. Improvements (i.e. seeding algorithm(s) to replace the current "triplet merging") are being worked on.