@slava77 2016-03-21
On 3/20/16 5:32 AM, Matti Kortelainen wrote:
>     initialStepSeeds* take about 6 times more (additional layer of
>     pixels isn't used effectively?) from 60 ms/ev to 370 ms/ev
> 
> Since you had "*" above I have to ask if you included
> |initialStepSeedsPreSplitting| there? For apples-to-apples comparison
> between trackingPhase1 and trackingPhase1PU70 only |initialStepSeeds|
> should be counted (pre-splitting step should accounted separately for
> the jet-core pixel cluster splitting). My tests show 2-2.5x increase in
> the timing of |initialStepSeeds|. This is caused mainly by the increase
> in the number of layer-triplets from 4 to 14 and that shows up in the
> timing. Improvements (i.e. seeding algorithm(s) to replace the current
> "triplet merging") are being worked on.
> 

From my job timing printouts:

self-increase	increase/job	baseline			new			name

   +1.451570      +0.40%        59.11 ms/ev ->       372.01 ms/ev
initialStepSeedsPreSplitting

   +1.453404      +0.40%        58.88 ms/ev ->       372.01 ms/ev
initialStepSeeds


They both change by about the same factor.

Thanks for all other explanations.